% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/france.trade.R
\docType{data}
\name{france.trade}
\alias{france.trade}
\title{France Trading Partners}
\usage{
data("france.trade")
}
\description{
Demo data set containing the trading percentage between France and different countries.
}
