% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phone.call.R
\docType{data}
\name{phone.call}
\alias{phone.call}
\title{Phone Call Data}
\usage{
data("phone.call")
}
\description{
Demo data set containing the number of daily phone calls between presidents of EU countries.
}
